repair: extract sink_source_for_repair
