+++ configure.py
@@ -854,6 +854,7 @@ scylla_core = (['database.cc',
                 'partition_slice_builder.cc',
                 'init.cc',
                 'lister.cc',
+                'repair/sink_source_for_repair.cc',
                 'repair/repair.cc',
                 'repair/row_level.cc',
                 'exceptions/exceptions.cc',
